♻️ refactor(reaction): move CAction-related files into Action folder (#24)
