+++ Source/Portfolio/Action/CAction_ComboAttack.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "CoreMinimal.h"
-#include "Weapon/CAction.h"
+#include "Action/CAction.h"
 #include "CAction_ComboAttack.generated.h"
 
 UCLASS()
